Adding new data and continuing work on Flask route for data retrieval from db
