+++ config.env
@@ -1,0 +1,7 @@
+# Setting environment variables
+
+db_password = "OraBoot123$"
+db_username = "root"
+db_url = "sweetdealdataclass.cszynmfuuwym.us-west-1.rds.amazonaws.com"
+db_port = "5432"
+db_name = "sweetdealdataclass"
