+++ test_schema/schema.sql
@@ -18,3 +18,5 @@ SELECT * FROM test;
 
 ALTER TABLE test DROP COLUMN index;
 ALTER TABLE test RENAME COLUMN timestamp TO time_stamp;
+
+DROP TABLE IF EXISTS test;
